TT-131: move handling invitation logic to the separated controller
