+++ spec/controllers/v1/users/passwords_controller_spec.rb
@@ -1,10 +1,10 @@
 require 'rails_helper'
 
-RSpec.describe V1::PasswordsController, type: :controller do
+RSpec.describe V1::Users::PasswordsController, type: :controller do
 
   describe "GET #new" do
     it "should return error in case of empty passed email" do
-      get :new, params: { email: "", format: :json }
+      post :create, params: { email: "", format: :json }
       expect(response.body).to eq({ error: I18n.t("passwords.email_does_not_present") }.to_json)
       expect(response.status).to eq(400)
     end
@@ -17,105 +17,30 @@ RSpec.describe V1::PasswordsController, type: :controller do
         expect(UserMailer).to receive(:recovery_password_email).with(user) { mailer }
         expect(mailer).to receive(:deliver_now)
 
-        get :new, params: { email: user.email, format: :json }
+        post :create, params: { email: user.email, format: :json }
       end
 
       it "should return success status in json if user was found" do
-        get :new, params: { email: user.email, format: :json }
+        post :create, params: { email: user.email, format: :json }
         expect(response.body).to eq({ status: "ok" }.to_json)
       end
 
       it "should return 400 status if user was found" do
-        get :new, params: { email: user.email, format: :json }
+        post :create, params: { email: user.email, format: :json }
         expect(response.status).to eq(200)
       end
 
       it "should return error if user was not found" do
-        get :new, params: { email: "test@gmail.com", format: :json }
+        post :create, params: { email: "test@gmail.com", format: :json }
         expect(response.body).to eq({
           error: I18n.t("passwords.user_not_found")
         }.to_json)
       end
 
       it "should return 404 if user was not found" do
-        get :new, params: { email: "test@gmail.com", format: :json }
+        post :create, params: { email: "test@gmail.com", format: :json }
         expect(response.status).to eq(404)
       end
-    end
-  end
-
-  describe "POST #create" do
-    it "should return error in case of empty passed token" do
-      post :create, params: { token: "", name: "", password: "", format: :json }
-      expect(response.body).to eq({ error: I18n.t("passwords.token_does_not_present") }.to_json)
-      expect(response.status).to eq(400)
-    end
-
-    it "should return error in case of empty passed name" do
-      post :create, params: { token: "222", name: "", password: "", format: :json }
-      expect(response.body).to eq({ error: I18n.t("passwords.name_does_not_present") }.to_json)
-      expect(response.status).to eq(400)
-    end
-
-    it "should return error in case of empty passed password" do
-      post :create, params: { token: "222", name: "test", password: "", format: :json }
-      expect(response.body).to eq({ error: I18n.t("passwords.password_does_not_present") }.to_json)
-      expect(response.status).to eq(400)
-    end
-
-    context "all params are valid" do
-      let!(:user) { create(:user) }
-      let(:request_params) { {
-        token: "22222222",
-        name: "John Doe",
-        password: "8932479238",
-        format: :json
-      } }
-
-      context "user was found by decoded token" do
-        before do
-          allow(TokenCryptService).to receive(:decode).with("22222222") { user.email }
-        end
-
-        it "should try to search user by decoded email" do
-          expect(User).to receive(:find_by).with({ email: user.email }) { user }
-          post :create, params: request_params
-        end
-
-        it "should set name for user" do
-          allow(User).to receive(:find_by) { user }
-          post :create, params: request_params
-          expect(user.reload.name == 'John Doe').to be_truthy
-        end
-
-        it "should set password for user" do
-          allow(User).to receive(:find_by) { user }
-          post :create, params: request_params
-          expect(user.reload.password == '8932479238').to be_truthy
-        end
-
-        it "should return 200 status" do
-          allow(User).to receive(:find_by) { user }
-          post :create, params: request_params
-          expect(response.status).to eq(200)
-        end
-
-        it "should return 422 status if user's password wasn't updated" do
-          allow(User).to receive(:find_by) { user }
-          allow(user).to receive(:save!) { false }
-          post :create, params: request_params
-          expect(response.status).to eq(422)
-        end
-
-        it "should return error message if user's password wasn't updated" do
-          allow(User).to receive(:find_by) { user }
-          allow(user).to receive(:save!) { false }
-          user.errors.add(:base, "error")
-          post :create, params: request_params
-          expect(response.body).to eq({ error: ["error"] }.to_json)
-        end
-      end
-
     end
   end
 
@@ -132,23 +57,11 @@ RSpec.describe V1::PasswordsController, type: :controller do
       expect(response.status).to eq(400)
     end
 
-    it "should return error in case of not matching password and confirmed one" do
-      put :update, params: {
-        token: "2222",
-        password: "11111111",
-        confirm_password: "999999999",
-        format: :json
-      }
-      expect(response.body).to eq({ error: I18n.t("passwords.confirmation_password_does_not_match") }.to_json)
-      expect(response.status).to eq(400)
-    end
-
     context "all params are valid" do
       let!(:user) { create(:user) }
       let(:request_params) { {
         token: "22222222",
         password: "8932479238",
-        confirm_password: "8932479238",
         format: :json
       } }
